requirements.txt
Django==5.2.3
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0

